Remove: type.ts 파일 삭제
